Holy Orders Parser output re-generated without faulty Titles (with negative levy_size)
Note to DR: this erases your hand-made changes to the localization of the titles for the HO Grand Masters. This is only temporary. I will re-inject your titles through the script, adding a colun for them in the Matrix and in the Parser, so that they get generated as per your design.

!!! Please do not edit by hand sections generated by the Parser. They should be edited in the Matrix / Python script only so that we can use the Parser again and again to generate code as we tweak and update the mod.